refactor(CompanionPlantAcceptanceTest): 인수테스트 step으로 리팩토링
